Moving LanguageSelector which is only used on layouts to layouts/shared folder
